discussion-forum/classesment-backend/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
Django==2.1.7
django-celery==3.2.2
djangorestframework==3.9.2
kombu==3.0.37
pytz==2018.9
vine==1.2.0
